+++ front/src/modules/simulationResult/components/SpaceCurvesSlopes.tsx
@@ -3,7 +3,8 @@ import { useSelector } from 'react-redux';
 import * as d3 from 'd3';
 import { CgLoadbar } from 'react-icons/cg';
 
-import { CHART_AXES, type PositionScaleDomain } from 'modules/simulationResult/consts';
+import { CHART_AXES } from 'modules/simulationResult/consts';
+import type { PositionScaleDomain } from 'modules/simulationResult/types';
 import {
   defineLinear,
   interpolateOnPosition,
