+++ front/src/modules/simulationResult/types.ts
@@ -1,0 +1,10 @@
+export type PositionScaleDomain = {
+  initial: number[];
+  current: number[];
+  source?: 'SpeedSpaceChart' | 'SpaceCurvesSlopes';
+};
+
+export type TimeScaleDomain = {
+  range?: [Date, Date];
+  source?: 'SpaceTimeChart' | 'Timeline';
+};
